+++ R/call_aston.R
@@ -34,7 +34,8 @@ sp_converter <- function(path, format_out = c("matrix", "data.frame", "data.tabl
   if (data_format == "long"){
     x <- reshape_chrom(x, data_format = "long")
   }
-  x <- convert_chrom_format(x, format_out = format_out)
+  x <- convert_chrom_format(x, format_out = format_out,
+                            data_format = data_format)
   if (read_metadata){
     meta <- read_masshunter_metadata(path)
     x <- attach_metadata(x, meta, format_in = metadata_format,
@@ -85,7 +86,7 @@ uv_converter <- function(path, format_out = c("matrix","data.frame","data.table"
   if (data_format == "long"){
     x <- reshape_chrom(x, data_format = "long")
   }
-  x <- convert_chrom_format(x, format_out = format_out)
+  x <- convert_chrom_format(x, format_out = format_out, data_format)
   if (correction){
     # multiply by empirical correction value
     correction_value <- 0.9536743164062551070259132757200859487056732177734375
@@ -126,7 +127,7 @@ trace_converter <- function(path, format_out = c("matrix", "data.frame"),
   if (data_format == "long"){
     x <- reshape_chrom(x, data_format = "long")
   }
-  x <- convert_chrom_format(x, format_out = format_out)
+  x <- convert_chrom_format(x, format_out = format_out, data_format)
   x
 }