+++ R/read_chemstation_csv.R
@@ -27,14 +27,15 @@ read_chemstation_csv <- function(path, format_out = c("matrix", "data.frame",
                                  data_format = c("wide", "long"),
                                  read_metadata = TRUE){
   format_out <- check_format_out(format_out)
-  data_format <- check_data_format(data_format, format_out)
+  data_format <- check_data_format(data_format, format_out = format_out)
   data <- read.csv(path, row.names = 1, header = TRUE,
                 fileEncoding = "utf-16LE", check.names = FALSE)
 
   if (data_format == "long"){
     data <- reshape_chrom_long(data)
   }
-  data <- convert_chrom_format(data, format_out = format_out)
+  data <- convert_chrom_format(data, format_out = format_out,
+                               data_format = data_format)
   if (read_metadata){
     data <- attach_metadata_minimal(data, data_format = data_format,
                                     format_out = format_out,